Add and Edit Product from Admin Dashboard (#7)
### Changes
- Now you can add and edit product seamlessly
- Added CORS middleware but it's not secure yet

+++ model/template.go
@@ -5,7 +5,7 @@ import (
 )
 
 type Template struct {
-	ID        uint       `gorm:"primary_key" json:"id" example:"1"`
+	ID        int64      `gorm:"primary_key" json:"id" example:"1"`
 	CreatedAt time.Time  `gorm:"not null;default:CURRENT_TIMESTAMP" json:"created_at" example:""`
 	UpdatedAt time.Time  `json:"updated_at" example:""`
 	DeletedAt *time.Time `json:"deleted_at" example:""`
